read epsilon noise directly from settings instead of calculating it

+++ Source/Simulation/Simulation.h
@@ -72,7 +72,7 @@ public:
   // demographic noise
   BoolParameter* stochasticity;
   RandomGausGenerator * rgg;
-  double noiseEpsilon; // = 1. / sqrt(volume)
+  float noiseEpsilon; // = 1. / sqrt(volume)
   
   
 	EnumParameter* setCAC;
